+++ src/main/java/com/jstarcraft/rns/utility/Trigamma.java
@@ -26,7 +26,7 @@ class Trigamma {
         float y = 0F;
         float z = 0F;
 
-        if (Double.isInfinite(x) || Double.isNaN(x)) {
+        if (Float.isInfinite(x) || Float.isNaN(x)) {
             return 0F / 0F;
         }
